+++ FinBoxLending/init/FinBoxLending.swift
@@ -9,7 +9,9 @@ import Foundation
 
 public class FinBoxLending {
     
-    private init(customerID: String, apiKey: String, userToken: String, environment: String, creditLineAmount: Float?, creditLineTransactionID: String?, showToolBar: Bool, hidePoweredBy: Bool, dcEnabled: Bool) {
+    private init(customerID: String, apiKey: String, userToken: String, environment: String, creditLineAmount: Float?, creditLineTransactionID: String?,
+                 showToolBar: Bool, hidePoweredBy: Bool, dcEnabled: Bool, utmTerm: String?, utmSource: String?, utmContent: String?, utmMedium: String?,
+                 utmCampaign: String?, utmPartnerName: String?, utmPartnerMedium: String?, appsflyerId: String?, idfa: String?, advertisingId: String?) {
         
     }
     
@@ -26,11 +28,21 @@ public class FinBoxLending {
         private var creditLineAmount: Float?
         private var creditLineTransactionID: String?
         // TODO: Add toolbar config
-//        private var toolBarConfig: CheckThisOut?
+        //        private var toolBarConfig: CheckThisOut?
         private var showToolBar: Bool?
         private var hidePoweredBy: Bool?
         private var dcEnabled: Bool?
-    
+        private var utmTerm: String?
+        private var utmSource: String?
+        private var utmContent: String?
+        private var utmMedium: String?
+        private var utmCampaign: String?
+        private var utmPartnerName: String?
+        private var utmPartnerMedium: String?
+        private var appsflyerId: String?
+        private var idfa: String?
+        private var advertisingId: String?
+        
         public func customerID(id: String) -> Builder {
             self.customerID = id
             return self
@@ -76,6 +88,56 @@ public class FinBoxLending {
             return self
         }
         
+        public func utmTerm(utmTerm: String) -> Builder {
+            self.utmTerm = utmTerm
+            return self
+        }
+        
+        public func utmSource(utmSource: String) -> Builder {
+            self.utmSource = utmSource
+            return self
+        }
+        
+        public func utmMedium(utmMedium: String) -> Builder {
+            self.utmMedium = utmMedium
+            return self
+        }
+        
+        public func utmContent(utmContent: String) -> Builder {
+            self.utmContent = utmContent
+            return self
+        }
+        
+        public func utmCampaign(utmCampaign: String) -> Builder {
+            self.utmCampaign = utmCampaign
+            return self
+        }
+        
+        public func utmPartnerName(utmPartnerName: String) -> Builder {
+            self.utmPartnerName = utmPartnerName
+            return self
+        }
+        
+        public func utmPartnerMedium(utmPartnerMedium: String) -> Builder {
+            self.utmPartnerMedium = utmPartnerMedium
+            return self
+        }
+        
+        public func appsflyerId(appsflyerId: String) -> Builder {
+            self.appsflyerId = appsflyerId
+            return self
+        }
+        
+        public func idfa(idfa: String) -> Builder {
+            self.idfa = idfa
+            return self
+        }
+        
+        public func advertisingId(advertisingId: String) -> Builder {
+            self.advertisingId = advertisingId
+            return self
+        }
+        
         public func build() -> FinBoxLending {
             guard let key = self.apiKey else {
                 fatalError("API Key cannot be null")
@@ -100,12 +162,28 @@ public class FinBoxLending {
             let hidePB = self.hidePoweredBy ?? false
             let dcEnabled = self.dcEnabled ?? false
             
-            savePreferences(customerID: id, apiKey: key, userToken: token, environment: env, creditLineAmount: cLineAmt, creditLineTransactionID: cLineTranxID, showToolBar: showTB, hidePoweredBy: hidePB, dcEnabled: dcEnabled)
+            let utmTerm = self.utmTerm
+            let utmSource = self.utmSource
+            let utmContent = self.utmContent
+            let utmMedium = self.utmMedium
+            let utmCampaign = self.utmCampaign
+            let utmPartnerName = self.utmPartnerName
+            let utmPartnerMedium = self.utmPartnerMedium
+            let appsflyerId = self.appsflyerId
+            let idfa = self.idfa
+            let advertisingId = self.advertisingId
             
-            return FinBoxLending(customerID: id, apiKey: key, userToken: token, environment: env, creditLineAmount: cLineAmt, creditLineTransactionID: cLineTranxID, showToolBar: showTB, hidePoweredBy: hidePB, dcEnabled: dcEnabled)
+            
+            savePreferences(customerID: id, apiKey: key, userToken: token, environment: env, creditLineAmount: cLineAmt, creditLineTransactionID: cLineTranxID, showToolBar: showTB, hidePoweredBy: hidePB, dcEnabled: dcEnabled, utmTerm: utmTerm, utmSource: utmSource, utmContent: utmContent, utmMedium: utmMedium, utmCampaign: utmCampaign, utmPartnerName: utmPartnerName, utmPartnerMedium: utmPartnerMedium, appsflyerId: appsflyerId, idfa: idfa, advertisingId: advertisingId)
+            
+            return FinBoxLending(customerID: id, apiKey: key, userToken: token, environment: env, creditLineAmount: cLineAmt, creditLineTransactionID: cLineTranxID, showToolBar: showTB, hidePoweredBy: hidePB, dcEnabled: dcEnabled, utmTerm: utmTerm, utmSource: utmSource, utmContent: utmContent, utmMedium: utmMedium, utmCampaign: utmCampaign, utmPartnerName: utmPartnerName, utmPartnerMedium: utmPartnerMedium, appsflyerId: appsflyerId, idfa: idfa, advertisingId: advertisingId
+            )
         }
         
-        public func savePreferences(customerID: String, apiKey: String, userToken: String, environment: String, creditLineAmount: Float?, creditLineTransactionID: String?, showToolBar: Bool, hidePoweredBy: Bool, dcEnabled: Bool) {
+        public func savePreferences(customerID: String, apiKey: String, userToken: String, environment: String, creditLineAmount: Float?,
+                                    creditLineTransactionID: String?, showToolBar: Bool, hidePoweredBy: Bool, dcEnabled: Bool,
+                                    utmTerm: String?, utmSource: String?, utmContent: String?, utmMedium: String?, utmCampaign: String?,
+                                    utmPartnerName: String?, utmPartnerMedium: String?, appsflyerId: String?, idfa: String?, advertisingId: String?) {
             let userPrefs = FinBoxLendingPref()
             
             userPrefs.apiKey = apiKey
@@ -116,6 +194,16 @@ public class FinBoxLending {
             userPrefs.creditLineTransactionID = creditLineTransactionID
             userPrefs.showToolBar = showToolBar
             userPrefs.hidePoweredBy = hidePoweredBy
+            userPrefs.utmTerm = utmTerm
+            userPrefs.utmSource = utmSource
+            userPrefs.utmContent = utmContent
+            userPrefs.utmMedium = utmMedium
+            userPrefs.utmCampaign = utmCampaign
+            userPrefs.utmPartnerName = utmPartnerName
+            userPrefs.utmPartnerMedium = utmPartnerMedium
+            userPrefs.appsflyerId = appsflyerId
+            userPrefs.idfa = idfa
+            userPrefs.advertisingId = advertisingId
         }
     }