+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 #  FinBox Lending iOS SDK
 
+## 0.2.2
+
+Add ABCD Endpoints
+Add Campaign Params
+Restrict to 15.0 Swift
+
+
 ## 0.1.0
 
 Update Callbacks